autogen-multiagent/requirements.txt
python-dotenv
autogen-agentchat
autogen-ext[openai,docker]
rich